skillrule multi-tenant: associate skillrule tenant to first queue tenant

+++ alembic/versions/73889d1db031_queueskillrule_add_tenant_uuid.py
@@ -13,26 +13,40 @@ import sqlalchemy as sa
 revision = '73889d1db031'
 down_revision = '0aeb61795700'
 
-queueskillrule_tbl = sa.sql.table('queueskillrule')
+queueskillrule_tbl = sa.sql.table('queueskillrule', sa.sql.column('tenant_uuid'))
+queuefeatures_tbl = sa.sql.table('queuefeatures', sa.sql.column('tenant_uuid'))
 
 
-def delete_all():
-    query = queueskillrule_tbl.delete()
-    op.execute(query)
+def associate_tenants():
+    query = sa.sql.select([queuefeatures_tbl.c.tenant_uuid])
+    result = op.get_bind().execute(query).first()
+
+    if result:
+        tenant_uuid = result[0]
+        query = (
+            queueskillrule_tbl.update()
+            .values(tenant_uuid=tenant_uuid)
+        )
+        op.execute(query)
+    else:
+        query = queueskillrule_tbl.delete()
+        op.execute(query)
 
 
 def upgrade():
-    delete_all()
-
     op.add_column(
         'queueskillrule',
         sa.Column(
             'tenant_uuid',
             sa.String(36),
             sa.ForeignKey('tenant.uuid', ondelete='CASCADE'),
-            nullable=False
+            nullable=True
         )
     )
+
+    associate_tenants()
+
+    op.alter_column('queueskillrule', 'tenant_uuid', nullable=False)
 
 
 def downgrade():
